requirements.txt
Django==6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
whitenoise==6.7.0
Pillow==10.3.0